fix: filter out outdated heartbeat (#2303)
* fix: filter out outdated heartbeat, #1707

* feat: reorder handlers

* refactor: disableXXX to enableXXX

* feat: make full use of region leases to facilitate failover

* chore: minor refactor

* chore: by comment

* feat: logging on inactive/active

+++ src/meta-srv/src/lease.rs
@@ -24,10 +24,10 @@ use crate::keys::{LeaseKey, LeaseValue, DN_LEASE_PREFIX};
 pub async fn alive_datanodes(
     cluster_id: u64,
     meta_peer_client: &MetaPeerClientRef,
-    lease_secs: i64,
+    lease_secs: u64,
 ) -> Result<HashMap<LeaseKey, LeaseValue>> {
     let lease_filter = |_: &LeaseKey, v: &LeaseValue| {
-        time_util::current_time_millis() - v.timestamp_millis < lease_secs * 1000
+        ((time_util::current_time_millis() - v.timestamp_millis) as u64) < lease_secs * 1000
     };
 
     filter_datanodes(cluster_id, meta_peer_client, lease_filter).await
